Move the table to its own component
Eventually the rest of the .erb can be react-ified and use the Explain component
